% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mapping.R
\name{robust.percapita}
\alias{robust.percapita}
\title{Estimate robust per capita rates from a event count map (numerator)
and a reference population map (denominator); both need to be
SpatRaster grids with the same dimensions and covering the same extent.
Event count map can be generated using robust.grid or robust.quadcount
from a point set.}
\usage{
robust.percapita(numerator, denominator, bd = NULL, weighted = F)
}
\description{
Estimate robust per capita rates from a event count map (numerator)
and a reference population map (denominator); both need to be
SpatRaster grids with the same dimensions and covering the same extent.
Event count map can be generated using robust.grid or robust.quadcount
from a point set.
}
