% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mapping.R
\name{robust.timeslices}
\alias{robust.timeslices}
\title{Estimate robust time slices}
\usage{
robust.timeslices(
  point_set,
  aggregation_method = c("count", "density"),
  opt_granularity = NULL,
  random_samples = T,
  nsamples = 500,
  signif = 0.99,
  tradeoff_crit = c("product", "sum"),
  uniformity_method = c("Quadratcount", "Nearest-neighbor"),
  robustness_method = c("Poisson", "Binomial", "Resampling"),
  robustness_k = -3,
  verbose = F,
  my_scales = NULL,
  W = NULL
)
}
\description{
Estimate robust time slices
}
