% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mapping.R
\name{robust.timeslices}
\alias{robust.timeslices}
\title{Robust time slices from gridded data stack}
\usage{
robust.timeslices(observed_counts_timestamp, verbose = F)
}
\description{
Improves the rates per cells for each timeslice by considering information
from the other slices. For each location, model (auto.arima only for now)
is estimated for the corresponding time series. The predicted value at that
cell-time is assigned as the improved, more robust rate. More formal
validation is still required, but simulation tests I've ran show that it
does yield better estimates.
}
